Add NetBSD support
The particular NetBSD box uses a shell script for sudo that
doesn't pass quoted args correctly (which kitchen-ansible needs
to fire up the ansible run), so we get/configure real sudo
in ansible pre_install.
Pkgin is a yum/apt like package manager for the NetBSD pacakges,
so we install, configure and tell ansible to use it by default.
For some reason Dlep was failing to build with either the gcc that
ships as part of the NetBSD base system (4.8) or 4.9 from packages
(cc1plus would just terminate while building DataItem.cpp).
clang builds seem fine though so we're building with the latest
clang available in NetBSD packages.