Add supported Alpine releases 3.6 and 3.8
3.8's Ansible packages are now rebased on python3,
so we need to tell kitchen-ansible that python3 is
the interpreter.

3.6 only comes with ansible 2.3. A2.3's apk module
appears to have issues with some packages, as
playbook runs kept failing to install 'g++' (or
verify its installation if I manually installed it).
Presuming this is an issue with "+" not being
understood as a valid character.

To workaround this, I've added a presinstall scriptlet
to install latest ansible via pip. First we apk install
pip and all of the dependencies for ansible, then
pip-install ansible.